Add files via upload
For the Zkav Club Vibe Design Contest, I created a vibrant and energetic design concept that embodies the spirit of a lively club. The goal was to design an immersive experience that would engage club members and create an unforgettable atmosphere. My objective was to merge cutting-edge design elements with the dynamic energy of the club